added email verification and reset password

+++ src/app/api/categories/[id]/route.js
@@ -1,12 +1,20 @@
 import { NextResponse } from "next/server";
-import connectDB from "../../../../../utils/dbConnect.js";
+import connectDB from "../../../../utils/dbConnect.js";
 // import TransactionsModel from "../../../../../models/TransactionsModel.js";
 import CategoriesModel from "../../../../../models/CategoriesModel.js";
+import { sanitizeInput } from "../../../../utils/sanitizeInput.js";
+import mongoose from "mongoose";
 
 export async function GET(request, { params }) {
   const { id } = await params;
   try {
     await connectDB();
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      return NextResponse.json(
+        { message: "ID is not a Valid ObjectId" },
+        { status: 403 }
+      );
+    }
     const txn = await CategoriesModel.findById(id);
     return NextResponse.json(txn);
   } catch (error) {
@@ -18,10 +26,17 @@ export async function GET(request, { params }) {
 }
 
 export async function PUT(request, { params }) {
-    const { id } = await params;
+  const { id } = await params;
   try {
     await connectDB();
-    const data = await request.json();
+    const rawData = await request.json();
+    const data = sanitizeInput(rawData);
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      return NextResponse.json(
+        { message: "ID is not a Valid ObjectId" },
+        { status: 403 }
+      );
+    }
     const txn = await CategoriesModel.findByIdAndUpdate(id, data, {
       new: true,
     }).lean();
@@ -38,6 +53,12 @@ export async function DELETE(request, { params }) {
   try {
     await connectDB();
     // const data = await request.json();
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      return NextResponse.json(
+        { message: "ID is not a Valid ObjectId" },
+        { status: 403 }
+      );
+    }
     const txn = await CategoriesModel.findByIdAndDelete(id);
     return NextResponse.json(txn);
   } catch (error) {
